Refine open-loop metric utilities and fixes
