+++ planning/autoware_planning_data_analyzer/src/metrics/metric_utils.cpp
@@ -1,0 +1,153 @@
+// Copyright 2026 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "metric_utils.hpp"
+
+#include <autoware/lanelet2_utils/intersection.hpp>
+#include <autoware_utils_geometry/geometry.hpp>
+
+#include <boost/geometry.hpp>
+
+#include <lanelet2_core/utility/Utilities.h>
+#include <tf2/utils.h>
+
+#include <memory>
+#include <unordered_set>
+
+namespace autoware::planning_data_analyzer::metrics
+{
+
+namespace
+{
+
+void append_unique_lanelet(
+  const lanelet::ConstLanelet & lanelet, lanelet::ConstLanelets & lanelets,
+  std::unordered_set<lanelet::Id> & seen_ids)
+{
+  if (seen_ids.insert(lanelet.id()).second) {
+    lanelets.push_back(lanelet);
+  }
+}
+
+}  // namespace
+
+bool is_vehicle_info_valid(const autoware::vehicle_info_utils::VehicleInfo & vehicle_info)
+{
+  return vehicle_info.vehicle_length_m > 0.0 && vehicle_info.vehicle_width_m > 0.0;
+}
+
+double get_yaw(const geometry_msgs::msg::Quaternion & orientation)
+{
+  return tf2::getYaw(orientation);
+}
+
+autoware_utils_geometry::Polygon2d create_pose_footprint(
+  const geometry_msgs::msg::Pose & pose,
+  const autoware::vehicle_info_utils::VehicleInfo & vehicle_info)
+{
+  const auto local_footprint = vehicle_info.createFootprint(0.0);
+  autoware_utils_geometry::Polygon2d polygon;
+  polygon.outer() = autoware_utils_geometry::transform_vector(
+    local_footprint, autoware_utils_geometry::pose2transform(pose));
+  boost::geometry::correct(polygon);
+  return polygon;
+}
+
+std::optional<lanelet::ConstLanelet> find_reference_lanelet(
+  const geometry_msgs::msg::Pose & pose,
+  const std::shared_ptr<autoware::route_handler::RouteHandler> & route_handler)
+{
+  if (!route_handler || !route_handler->isHandlerReady()) {
+    return std::nullopt;
+  }
+
+  lanelet::ConstLanelet closest_lanelet;
+  if (route_handler->getClosestLaneletWithinRoute(pose, &closest_lanelet)) {
+    return closest_lanelet;
+  }
+
+  for (const auto & lanelet : route_handler->getRoadLaneletsAtPose(pose)) {
+    if (route_handler->isRouteLanelet(lanelet)) {
+      return lanelet;
+    }
+  }
+
+  return std::nullopt;
+}
+
+lanelet::ConstLanelets collect_route_relevant_lanelets(
+  const autoware_planning_msgs::msg::Trajectory & trajectory,
+  const std::shared_ptr<autoware::route_handler::RouteHandler> & route_handler)
+{
+  lanelet::ConstLanelets route_lanelets;
+  if (!route_handler || !route_handler->isHandlerReady()) {
+    return route_lanelets;
+  }
+
+  std::unordered_set<lanelet::Id> seen_ids;
+  for (const auto & point : trajectory.points) {
+    for (const auto & lanelet : route_handler->getRoadLaneletsAtPose(point.pose)) {
+      if (!route_handler->isRouteLanelet(lanelet)) {
+        continue;
+      }
+      append_unique_lanelet(lanelet, route_lanelets, seen_ids);
+    }
+  }
+
+  return route_lanelets;
+}
+
+autoware_utils_geometry::LineString2d to_linestring2d(const lanelet::ConstLineString3d & line)
+{
+  autoware_utils_geometry::LineString2d line_2d;
+  for (const auto & point : lanelet::utils::to2D(line)) {
+    line_2d.push_back({point.x(), point.y()});
+  }
+  return line_2d;
+}
+
+bool is_pose_in_intersection(
+  const geometry_msgs::msg::Pose & pose,
+  const std::shared_ptr<autoware::route_handler::RouteHandler> & route_handler)
+{
+  const auto lanelet = find_reference_lanelet(pose, route_handler);
+  return lanelet.has_value() &&
+         autoware::experimental::lanelet2_utils::is_intersection_lanelet(*lanelet);
+}
+
+double forward_offset_in_ego_frame(
+  const geometry_msgs::msg::Pose & ego_pose, const geometry_msgs::msg::Pose & object_pose)
+{
+  const double yaw = get_yaw(ego_pose.orientation);
+  const double dx = object_pose.position.x - ego_pose.position.x;
+  const double dy = object_pose.position.y - ego_pose.position.y;
+  return std::cos(yaw) * dx + std::sin(yaw) * dy;
+}
+
+bool is_agent_behind(
+  const geometry_msgs::msg::Pose & ego_pose, const geometry_msgs::msg::Pose & object_pose)
+{
+  return forward_offset_in_ego_frame(ego_pose, object_pose) < 0.0;
+}
+
+const autoware_perception_msgs::msg::PredictedPath * highest_confidence_path(
+  const autoware_perception_msgs::msg::PredictedObject & object)
+{
+  const auto it = std::max_element(
+    object.kinematics.predicted_paths.begin(), object.kinematics.predicted_paths.end(),
+    [](const auto & lhs, const auto & rhs) { return lhs.confidence < rhs.confidence; });
+  return it == object.kinematics.predicted_paths.end() ? nullptr : &(*it);
+}
+
+}  // namespace autoware::planning_data_analyzer::metrics
