+++ planning/autoware_planning_data_analyzer/src/metrics/ttc_within_bound.cpp
@@ -14,15 +14,14 @@
 
 #include "ttc_within_bound.hpp"
 
-#include <autoware/lanelet2_utils/intersection.hpp>
+#include "metric_utils.hpp"
+
 #include <autoware_utils_geometry/boost_geometry.hpp>
 #include <autoware_utils_geometry/boost_polygon_utils.hpp>
 #include <autoware_utils_geometry/geometry.hpp>
 #include <tf2/LinearMath/Vector3.hpp>
 
 #include <boost/geometry.hpp>
-
-#include <tf2/utils.h>
 
 #include <algorithm>
 #include <array>
@@ -31,8 +30,6 @@
 #include <memory>
 #include <optional>
 #include <string>
-#include <unordered_set>
-
 namespace autoware::planning_data_analyzer::metrics
 {
 
@@ -45,16 +42,6 @@ using autoware_utils_geometry::Polygon2d;
 
 constexpr double kStoppedSpeedThreshold = 5.0e-3;
 constexpr std::array<double, 4> kFutureProjectionOffsetsSec{{0.0, 0.3, 0.6, 0.9}};
-
-bool is_vehicle_info_valid(const autoware::vehicle_info_utils::VehicleInfo & vehicle_info)
-{
-  return vehicle_info.vehicle_length_m > 0.0 && vehicle_info.vehicle_width_m > 0.0;
-}
-
-double get_yaw(const geometry_msgs::msg::Quaternion & orientation)
-{
-  return tf2::getYaw(orientation);
-}
 
 tf2::Vector3 get_velocity_in_world_coordinate(
   const autoware_planning_msgs::msg::TrajectoryPoint & point)
@@ -76,77 +63,10 @@ geometry_msgs::msg::Pose project_pose(
   return projected;
 }
 
-Polygon2d create_pose_footprint(
-  const geometry_msgs::msg::Pose & pose,
-  const autoware::vehicle_info_utils::VehicleInfo & vehicle_info)
-{
-  const auto local_footprint = vehicle_info.createFootprint(0.0);
-  Polygon2d polygon;
-  polygon.outer() = autoware_utils_geometry::transform_vector(
-    local_footprint, autoware_utils_geometry::pose2transform(pose));
-  boost::geometry::correct(polygon);
-  return polygon;
-}
-
-std::optional<lanelet::ConstLanelet> find_reference_lanelet(
-  const geometry_msgs::msg::Pose & pose,
-  const std::shared_ptr<autoware::route_handler::RouteHandler> & route_handler)
-{
-  if (!route_handler || !route_handler->isHandlerReady()) {
-    return std::nullopt;
-  }
-
-  lanelet::ConstLanelet closest_lanelet;
-  if (route_handler->getClosestLaneletWithinRoute(pose, &closest_lanelet)) {
-    return closest_lanelet;
-  }
-
-  for (const auto & lanelet : route_handler->getRoadLaneletsAtPose(pose)) {
-    if (route_handler->isRouteLanelet(lanelet)) {
-      return lanelet;
-    }
-  }
-
-  return std::nullopt;
-}
-
-bool is_pose_in_intersection(
-  const geometry_msgs::msg::Pose & pose,
-  const std::shared_ptr<autoware::route_handler::RouteHandler> & route_handler)
-{
-  const auto lanelet = find_reference_lanelet(pose, route_handler);
-  return lanelet.has_value() &&
-         autoware::experimental::lanelet2_utils::is_intersection_lanelet(*lanelet);
-}
-
-double forward_offset_in_ego_frame(
-  const geometry_msgs::msg::Pose & ego_pose, const geometry_msgs::msg::Pose & object_pose)
-{
-  const double yaw = get_yaw(ego_pose.orientation);
-  const double dx = object_pose.position.x - ego_pose.position.x;
-  const double dy = object_pose.position.y - ego_pose.position.y;
-  return std::cos(yaw) * dx + std::sin(yaw) * dy;
-}
-
 bool is_agent_ahead(
   const geometry_msgs::msg::Pose & ego_pose, const geometry_msgs::msg::Pose & object_pose)
 {
   return forward_offset_in_ego_frame(ego_pose, object_pose) > 0.0;
-}
-
-bool is_agent_behind(
-  const geometry_msgs::msg::Pose & ego_pose, const geometry_msgs::msg::Pose & object_pose)
-{
-  return forward_offset_in_ego_frame(ego_pose, object_pose) < 0.0;
-}
-
-const autoware_perception_msgs::msg::PredictedPath * highest_confidence_path(
-  const autoware_perception_msgs::msg::PredictedObject & object)
-{
-  const auto it = std::max_element(
-    object.kinematics.predicted_paths.begin(), object.kinematics.predicted_paths.end(),
-    [](const auto & lhs, const auto & rhs) { return lhs.confidence < rhs.confidence; });
-  return it == object.kinematics.predicted_paths.end() ? nullptr : &(*it);
 }
 
 std::optional<geometry_msgs::msg::Pose> interpolate_object_pose(
@@ -210,7 +130,6 @@ TTCWithinBoundResult calculate_ttc_within_bound(
     return result;
   }
 
-  std::unordered_set<std::size_t> collided_object_indices;
   for (const auto & point : trajectory.points) {
     const auto velocity_world = get_velocity_in_world_coordinate(point);
     const double speed = std::hypot(velocity_world.x(), velocity_world.y());
@@ -227,10 +146,6 @@ TTCWithinBoundResult calculate_ttc_within_bound(
       const auto ego_polygon = create_pose_footprint(projected_pose, vehicle_info);
 
       for (std::size_t object_idx = 0; object_idx < objects->objects.size(); ++object_idx) {
-        if (collided_object_indices.count(object_idx) != 0U) {
-          continue;
-        }
-
         const auto & object = objects->objects.at(object_idx);
         const auto object_pose = interpolate_object_pose(object, query_time_s);
         if (!object_pose.has_value()) {
@@ -248,11 +163,9 @@ TTCWithinBoundResult calculate_ttc_within_bound(
           (ego_in_intersection && !is_agent_behind(point.pose, *object_pose))) {
           result.score = 0.0;
           result.reason = "collision_within_bound";
-          result.infraction_time_s = rclcpp::Duration(point.time_from_start).seconds();
+          result.infraction_time_s = query_time_s;
           return result;
         }
-
-        collided_object_indices.insert(object_idx);
       }
     }
   }
